requirements.txt
openai>=1.3.5
gspread>=5.11.0
google-auth>=2.22.0
twilio>=8.10.0
python-dotenv>=1.0.0
